+++ Proyecto/Movie.h
@@ -9,23 +9,23 @@ private:
 	std::string movieCountry;
 	int movieYear;
 	int movieLength;
-	int movieReview;
+	float movieReview;
 
 public:
 	Movie();
 	Movie(std::string aMovieName, std::string aMovieCountry,
-		int aMovieYear, int aMovieLength, int aMovieReview);
+		int aMovieYear, int aMovieLength, float aMovieReview);
 
 	void setMovieName(std::string aMovieName);
 	void setMovieCountry(std::string aMovieCountry);
 	void setMovieYear(int aMovieYear);
 	void setMovieLength(int aMovieLength);
-	void setMovieReview(int aMovieReview);
+	void setMovieReview(float aMovieReview);
 
 	std::string getMovieName();
 	std::string getMovieCountry();
 	int getMovieYear();
 	int getMovieLength();
-	int getMovieReview();
+	float getMovieReview();
 };
 #endif
